requirements.txt
GitPython==3.1.40
pytest==7.4.3
pytest_cov==4.1.0
pytest_json_report==1.5.0
pythonbasictools==0.0.1a10
PyGithub~=2.1.1